Menu Updated
There was an issue where the customer_id wasn’t being picked up when a user created a new account and placed a pizza order. Now, after successfully creating an account, the login and password are shown right away. This way, when the user logs in, the pizza order can correctly get the customer_id from the login process.